+++ cl-syswork20r/src/components/Forms/AddPersonForm.css
@@ -9,7 +9,14 @@
 
 .color-button-edit {
     /* background: #416ce0 linear-gradient(180deg, #3b76c4, #416ce0) repeat-x; */
-    background: #362d86 linear-gradient(270deg, #483797, #362d86) repeat-x;
+    /* background: #362d86 linear-gradient(270deg, #483797, #362d86) repeat-x; */
+    background: #362d86 linear-gradient(270deg, #7e21d4, #362d86) repeat-x;
+    color: rgba(231, 231, 240, 0.952);
+}
+
+.color-button-edit-employee{
+    /* background: #416ce0 linear-gradient(180deg, #3b76c4, #416ce0) repeat-x; */
+    background: #1b53a8 linear-gradient(270deg, #1695cf, #1b53a8) repeat-x;
     color: rgba(231, 231, 240, 0.952);
 }
