+++ srv-syswork20r/src/management/querys/accounting.js
@@ -1,21 +1,20 @@
 const queryAccountingModel = {
     
-    currency:{
-        getListCurrencies: (data)=>{
-            return `SELECT 
+    getListCurrencies: (data)=>{
+        return `SELECT 
 
-            CRR.idcurrency,
-            CRR.currency,
-            CRR.symbol,
-            CRR.is_enabled,
-            IF(CRR.is_enabled = 1, 'ENABLED', 'DISABLED') AS state,
-            CRR.description
-            
-            FROM currency as CRR 
-            
-            WHERE ${data.filter}`
-        }
+        CRR.idcurrency,
+        CRR.currency,
+        CRR.symbol,
+        CRR.is_enabled,
+        IF(CRR.is_enabled = 1, 'ENABLED', 'DISABLED') AS state,
+        CRR.description
+        
+        FROM currency as CRR 
+        
+        WHERE ${data.filter}`
     }
+    
 }
